+++ AdventureBackpacks/AdventureBackpacks.cs
@@ -40,7 +40,7 @@ namespace AdventureBackpacks
         //Module Constants
         private const string _pluginId = "vapok.mods.adventurebackpacks";
         private const string _displayName = "Adventure Backpacks";
-        private const string _version = "2.0.10";
+        private const string _version = "2.1.0";
 
         //Interface Properties
         public string PluginId => _pluginId;
@@ -126,6 +126,8 @@ namespace AdventureBackpacks
 
             //Initialized Features
             QuickTransfer.FeatureInitialized = true;
+            CraftFromBackpack.FeatureInitialized = true;
+            StoreToBackpack.FeatureInitialized = true;
         }
 
         private void Update()